+++ src/dataflow-types/src/lib.rs
@@ -14,6 +14,7 @@
 //! avoid the dependency, as the dataflow crate is very slow to compile.
 
 pub mod logging;
+pub mod plan;
 
 mod errors;
 mod explain;
@@ -23,4 +24,5 @@ pub use errors::*;
 pub use explain::DataflowGraphFormatter;
 pub use explain::Explanation;
 pub use explain::JsonViewFormatter;
+pub use plan::Plan;
 pub use types::*;
