+++ src/dataflow/src/logging/materialized.rs
@@ -25,8 +25,8 @@ use super::{LogVariant, MaterializedLog};
 use crate::activator::RcActivator;
 use crate::arrangement::manager::RowSpine;
 use crate::arrangement::KeysValsHandle;
-use crate::render::Permutation;
 use crate::replay::MzReplay;
+use dataflow_types::plan::Permutation;
 use expr::{GlobalId, SourceInstanceId};
 use repr::adt::jsonb::Jsonb;
 use repr::{Datum, DatumVec, Row, Timestamp};
