+++ src/dataflow/src/logging/reachability.rs
@@ -22,9 +22,9 @@ use crate::activator::RcActivator;
 use crate::arrangement::manager::RowSpine;
 use crate::arrangement::KeysValsHandle;
 use crate::logging::ConsolidateBuffer;
-use crate::render::Permutation;
 use crate::replay::MzReplay;
 use dataflow_types::logging::LoggingConfig;
+use dataflow_types::plan::Permutation;
 use ore::iter::IteratorExt;
 use repr::{Datum, Row, RowArena, Timestamp};
